tflite: added missing allocate_tensors call
as well as some apparently needed shared libraries for linux

+++ eye-ai-core-rs/build.rs
@@ -9,46 +9,23 @@ fn main() {
 
 	// copying .so files from third_party
 	let target_os = std::env::var("CARGO_CFG_TARGET_OS").unwrap();
-	let out_dir = std::env::var("OUT_DIR").unwrap();
 	let third_party_dir = format!("third_party/{}", target_os);
+	println!("cargo::rerun-if-changed={}", third_party_dir);
+	let out_dir = std::env::var("OUT_DIR").unwrap();
+	let target_output_dir = format!("{}/../../../", out_dir);
 
-	// copy tensorflowlite .so file
-	let libtensorflowlite_so_filename = if target_os == "android" {
-		"libtensorflowlite_jni.so"
-	} else {
-		"libtensorflow-lite.so"
-	};
-	let libtensorflowlite_so = format!("{}/{}", third_party_dir, libtensorflowlite_so_filename);
-	println!("cargo::rerun-if-changed={}", libtensorflowlite_so);
-	std::fs::copy(
-		libtensorflowlite_so,
-		format!("{}/../../../{}", out_dir, libtensorflowlite_so_filename),
-	)
-	.expect("failed to copy tensorflow-lite library file");
-
-	if target_os == "linux" {
-		// copy libabsl_log_internal_nullguard.so
-		let libabsl_log_internal_nullguard_so =
-			format!("{}/libabsl_log_internal_nullguard.so", third_party_dir);
-		println!(
-			"cargo::rerun-if-changed={}",
-			libabsl_log_internal_nullguard_so
-		);
-		std::fs::copy(
-			libabsl_log_internal_nullguard_so,
-			format!("{}/../../../libabsl_log_internal_nullguard.so", out_dir),
-		)
-		.expect("failed to copy absl log internal nullguard library file");
-	}
-
-	// copy libqnn_delegate_jni.so
-	if target_os == "android" {
-		let qnn_delegate_library_so = format!("{}/libqnn_delegate_jni.so", third_party_dir);
-		println!("cargo::rerun-if-changed={}", qnn_delegate_library_so);
-		std::fs::copy(
-			qnn_delegate_library_so,
-			format!("{}/../../../libqnn_delegate_jni.so", out_dir),
-		)
-		.expect("failed to copy qnn delegate library file");
+	for entry in std::fs::read_dir(&third_party_dir)
+		.expect("failed to walk files of third_pary_dir")
+		.flatten()
+	{
+		if entry.file_type().unwrap().is_file() {
+			let path = entry.path();
+			let filename = path.file_name().unwrap();
+			std::fs::copy(
+				entry.path(),
+				format!("{}/{}", target_output_dir, filename.to_str().unwrap()),
+			)
+			.expect("failed to copy .so file to target output directory");
+		}
 	}
 }
